Add support for exclusion based on row or column restrictions detected in a block.
If the 2 or 3 possible positions for a value in a block occur in a row
(or column) then the values cannot occur in any other cell in the
intersecting row.

We use a cheap test to schedule some expensive work which we will only
use if we run out of other things to try.

We take advantage of the fact that cells are stored in sorted order to
exclude from consideration those cells in the intersection itself.

We also avoid the rejections if the count of cells found in the intersection
isn't identical to the possible positions in the block.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>