Add backtracking support and also ambiguous puzzle detection.
When we run out of heuristics to try, we try to brute
force a solution by cloning the current grid and
asserting one of the possible values that has yet to be decided (this is a guess).

If this works, then we check that there is no other solution with
that value rejected. If there is a solution with a different value in the same
cell then the puzzle is ambiguous and hence invalid. If not, the solution
is of the cloned puzzle is unique and we copy it into the original puzzle and
declare success.

If assertion of the guessed value fails, we reject that value in the main
puzzle and continue with the heuristic search falling back to backtracking again as
heuristics are exhausted.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>